Create UserConnector (#2723)
Merge UserAccountConnector and UserDetailsConnector into UserConnector including related adapters and converters

Closes GH-2722
