+++ projects/core/src/occ/adapters/user/user-occ.module.ts
@@ -4,12 +4,10 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { UserOrderAdapter } from '../../../user/connectors/order/user-order.adapter';
 import { OccUserOrderAdapter } from './occ-user-order.adapter';
-import { UserDetailsAdapter } from '../../../user/connectors/details/user-details.adapter';
-import { OccUserDetailsAdapter } from './occ-user-details.adapter';
 import { UserAddressAdapter } from '../../../user/connectors/address/user-address.adapter';
 import { OccUserAddressAdapter } from './occ-user-address.adapter';
-import { UserAccountAdapter } from '../../../user/connectors/account/user-account.adapter';
-import { OccUserAccountAdapter } from './occ-user-account.adapter';
+import { UserAdapter } from '../../../user/connectors/user/user.adapter';
+import { OccUserAdapter } from './occ-user.adapter';
 import { UserPaymentAdapter } from '../../../user/connectors/payment/user-payment.adapter';
 import { OccUserPaymentAdapter } from './occ-user-payment.adapter';
 import { OccUserConsentAdapter } from './occ-user-consent.adapter';
@@ -18,9 +16,8 @@ import { UserConsentAdapter } from '../../../user/connectors/consent/user-consen
 @NgModule({
   imports: [CommonModule, HttpClientModule],
   providers: [
-    { provide: UserDetailsAdapter, useClass: OccUserDetailsAdapter },
+    { provide: UserAdapter, useClass: OccUserAdapter },
     { provide: UserAddressAdapter, useClass: OccUserAddressAdapter },
-    { provide: UserAccountAdapter, useClass: OccUserAccountAdapter },
     { provide: UserConsentAdapter, useClass: OccUserConsentAdapter },
     {
       provide: UserPaymentAdapter,
