--- projects/core/src/user/connectors/account/index.ts
@@ -1,3 +1,0 @@
-export * from './user-account.connector';
-export * from './user-account.adapter';
-export * from './converters';
