+++ projects/core/src/user/connectors/index.ts
@@ -1,6 +1,5 @@
-export * from './account/index';
+export * from './user/index';
 export * from './address/index';
 export * from './consent/index';
-export * from './details/index';
 export * from './payment/index';
 export * from './order/index';
