+++ projects/core/src/user/connectors/user/index.ts
@@ -1,0 +1,3 @@
+export * from './user.connector';
+export * from './user.adapter';
+export * from './converters';
