+++ projects/core/src/user/connectors/user/user.adapter.ts
@@ -1,7 +1,11 @@
 import { Observable } from 'rxjs';
 import { Title, User, UserSignUp } from '../../../model/misc.model';
 
-export abstract class UserAccountAdapter {
+export abstract class UserAdapter {
+  abstract load(userId: string): Observable<User>;
+
+  abstract update(username: string, user: User): Observable<{}>;
+
   abstract register(user: UserSignUp): Observable<User>;
 
   abstract requestForgotPasswordEmail(userEmailAddress: string): Observable<{}>;
