+++ projects/core/src/user/store/effects/forgot-password.effect.ts
@@ -7,7 +7,7 @@ import { catchError, concatMap, map, switchMap } from 'rxjs/operators';
 import { AddMessage, GlobalMessageType } from '../../../global-message/index';
 
 import * as fromActions from '../actions/index';
-import { UserAccountConnector } from '../../connectors/account/user-account.connector';
+import { UserConnector } from '../../connectors/user/user.connector';
 
 @Injectable()
 export class ForgotPasswordEffects {
@@ -41,6 +41,6 @@ export class ForgotPasswordEffects {
 
   constructor(
     private actions$: Actions,
-    private userAccountConnector: UserAccountConnector
+    private userAccountConnector: UserConnector
   ) {}
 }
