+++ projects/core/src/user/store/effects/user-details.effect.ts
@@ -4,7 +4,7 @@ import { Observable, of } from 'rxjs';
 import { catchError, concatMap, map, mergeMap } from 'rxjs/operators';
 import * as fromUserDetailsAction from '../actions/user-details.action';
 import { User } from '../../../model/misc.model';
-import { UserDetailsConnector } from '../../connectors/details/user-details.connector';
+import { UserConnector } from '../../connectors/user/user.connector';
 
 @Injectable()
 export class UserDetailsEffects {
@@ -15,7 +15,7 @@ export class UserDetailsEffects {
     ofType(fromUserDetailsAction.LOAD_USER_DETAILS),
     map((action: fromUserDetailsAction.LoadUserDetails) => action.payload),
     mergeMap(userId => {
-      return this.userDetailsConnector.get(userId).pipe(
+      return this.userConnector.get(userId).pipe(
         map((user: User) => {
           return new fromUserDetailsAction.LoadUserDetailsSuccess(user);
         }),
@@ -34,24 +34,22 @@ export class UserDetailsEffects {
     ofType(fromUserDetailsAction.UPDATE_USER_DETAILS),
     map((action: fromUserDetailsAction.UpdateUserDetails) => action.payload),
     concatMap(payload =>
-      this.userDetailsConnector
-        .update(payload.username, payload.userDetails)
-        .pipe(
-          map(
-            _ =>
-              new fromUserDetailsAction.UpdateUserDetailsSuccess(
-                payload.userDetails
-              )
-          ),
-          catchError(error =>
-            of(new fromUserDetailsAction.UpdateUserDetailsFail(error))
-          )
+      this.userConnector.update(payload.username, payload.userDetails).pipe(
+        map(
+          _ =>
+            new fromUserDetailsAction.UpdateUserDetailsSuccess(
+              payload.userDetails
+            )
+        ),
+        catchError(error =>
+          of(new fromUserDetailsAction.UpdateUserDetailsFail(error))
         )
+      )
     )
   );
 
   constructor(
     private actions$: Actions,
-    private userDetailsConnector: UserDetailsConnector
+    private userConnector: UserConnector
   ) {}
 }
